requirements.txt
blinker==1.8.2
click==8.1.7
Flask==3.0.3
gunicorn==20.1.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
setuptools==72.2.0
Werkzeug==3.0.3
tensorflow-cpu==2.17.0
Pillow==9.2.0
